% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_tags.R
\name{edit_tags}
\alias{edit_tags}
\title{Shiny app for batch tag editing}
\usage{
edit_tags(path = getwd())
}
\arguments{
\item{path}{path to where the posts are found}
}
\details{
The YAML metadata can also be edited in other fields, e.g. a slug
can become lala instead of 'lala' which should not pose any problem.
That said, do check all edits before committing them (assuming your website is
 under version control!).
}
