% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Function returning tags}
\usage{
get_tags(path, format)
}
\arguments{
\item{path}{Path to where the posts are located}

\item{format}{String of post formats}
}
\value{
data.frame of filenames and tags as Booleans
}
\description{
It returns a data.frame of filenames and tags as Booleans. A nice rectangle!
}
\examples{
blog_path <- system.file(package = "blogyaml", "example_blog")
get_tags(blog_path, format = "*.md")
}
