% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_tags.R
\name{inject_tags}
\alias{inject_tags}
\title{Title}
\usage{
inject_tags(path, tags)
}
\arguments{
\item{path}{Path to where the posts are located}

\item{tags}{Tags as returned by \code{get_tags} and then edited}
}
\description{
Title
}
\examples{
\dontrun{
path <- system.file(package = "blogyaml", "example_blog")
tags <- get_tags(path, format = "*.md")
# do something to tags
inject_tags(path, tags)
}
}
