% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags.R
\name{get_tags}
\alias{get_tags}
\title{Title}
\usage{
get_tags(path)
}
\arguments{
\item{path}{Path to where the posts are located}
}
\value{
data.frame of filenames and tags as Booleans
}
\description{
Title
}
